// src/orm/orm.enum.ts
export enum OrmQ {
  SELECT = "SELECT",
  ENDL = "\n",
  ALL = "*",
  FROM = "FROM",
  WHERE = "WHERE",
  JOIN = "JOIN",
  JOIN_ON = "ON",
  EQUALS = "=",
  IN_OPERATOR = "IN",
  OR_OPERATOR = "OR",
  INSERT = "INSERT",
  INSERT_VALUES = "VALUES",
  ORDER_BY = "ORDER BY",
  GROUP_BY = "GROUP BY",
}